Imported Upstream version 5.0.3

+++ src/cluster/smoothest.cc
@@ -172,14 +172,14 @@ unsigned long standardise(volume<float>& mask,
 	    double Sx = 0.0, SSx = 0.0;
 	    
 	    for ( int t = 0; t < M; t++ ) {
-	      float R_it = R(x,y,z,t);
+	      double R_it = R(x,y,z,t);
 	      
 	      Sx += R_it;
 	      SSx += Sqr(R_it);
 	    }
 	    
-	    float mean = Sx / M;
-	    float sdsq = (SSx - (Sqr(Sx) / M)) / (M - 1) ;
+	    double mean = Sx / M;
+	    double sdsq = (SSx - (Sqr(Sx) / M)) / (M - 1) ;
 	    
 	    if (sdsq<=0) {
 	      // trap for differences between mask and invalid data
@@ -299,7 +299,7 @@ int main(int argc, char **argv) {
   // Estimate the smoothness of the normalised residual field
   // see TR00DF1 for mathematical description of the algorithm.
   enum {X = 0, Y, Z};
-  float SSminus[3] = {0, 0, 0}, S2[3] = {0, 0, 0};
+  double SSminus[3] = {0, 0, 0}, S2[3] = {0, 0, 0};
 
   int zstart=1;
   if (!usez) zstart=0;
@@ -326,8 +326,8 @@ int main(int argc, char **argv) {
 	  }
 	}
 
-  float norm = 1.0/(float) N;
-  float v = dof.value();	// v - degrees of freedom (nu)  
+  double norm = 1.0/(double) N;
+  double v = dof.value();	// v - degrees of freedom (nu)  
   if(R.tsize() > 1) {
     if(verbose.value()) {
       cerr << "Non-edge voxels = " << N << endl;
@@ -367,7 +367,7 @@ int main(int argc, char **argv) {
   }
 
   // Convert to sigma squared
-  float sigmasq[3];
+  double sigmasq[3];
 
   sigmasq[X] = -1.0 / (4 * log(fabs(SSminus[X]/S2[X])));
   sigmasq[Y] = -1.0 / (4 * log(fabs(SSminus[Y]/S2[Y])));
@@ -378,7 +378,7 @@ int main(int argc, char **argv) {
   // Furthermore, W_i = 1/(2.lambda_i) = sigma_i^2 => 
   //   det(Lambda) = det( lambda_i ) = det ( (2 W_i)^-1 ) = (2^D det(W))^-1
   //   where D = number of dimensions (2 or 3)
-  float dLh;
+  double dLh;
   if (usez) { dLh=pow(sigmasq[X]*sigmasq[Y]*sigmasq[Z], -0.5)*pow(8, -0.5); }
   else { dLh = pow(sigmasq[X]*sigmasq[Y], -0.5)*pow(4, -0.5); }
 
@@ -388,12 +388,12 @@ int main(int argc, char **argv) {
   if(R.tsize() > 1) dLh *= SMOOTHEST::interpolate(v);
 
   // Convert to full width half maximum
-  float FWHM[3];
+  double FWHM[3];
   FWHM[X] = sqrt(8 * log(2) * sigmasq[X]);
   FWHM[Y] = sqrt(8 * log(2) * sigmasq[Y]);
   if (usez) { FWHM[Z] = sqrt(8 * log(2) * sigmasq[Z]); }
   else { FWHM[Z]=0; }
-  float resels = FWHM[X] * FWHM[Y];
+  double resels = FWHM[X] * FWHM[Y];
   if (usez) resels *= FWHM[Z];
 
   if(verbose.value()) {
